add: 可忽略数据同步时的错误 fix: 渲染装备图时找不到文件的bug 第一次启动报错缺少error.log

+++ AZbot/plugins/checker/start_checker.py
@@ -19,8 +19,9 @@ async def _(bot: Bot):
 
 @driver.on_startup
 async def init():
-    shutil.copyfile("error.log", "error.log.bak")
-    os.remove("error.log")
+    if os.path.exists("error.log"):
+        shutil.copyfile("error.log", "error.log.bak")
+        os.remove("error.log")
     logger.add("error.log", level="ERROR", format=default_format)
     # 文件检查
     proxy = cfg["base"]["network_proxy"]
